Move the core::fmt::Write impl into the zerotrie crate
