+++ provider/blob/src/blob_schema.rs
@@ -3,11 +3,10 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 use alloc::boxed::Box;
-use core::fmt;
 use icu_provider::prelude::*;
 use serde::Deserialize;
 use writeable::Writeable;
-use zerotrie::{ZeroTrieSimpleAscii, ZeroTrieSimpleAsciiCursor};
+use zerotrie::ZeroTrieSimpleAscii;
 use zerovec::maps::{ZeroMap2dBorrowed, ZeroMapKV};
 use zerovec::vecs::{Index32, VarZeroSlice, VarZeroVec, ZeroSlice};
 
@@ -150,33 +149,6 @@ impl Default for BlobSchemaV2<'_> {
     }
 }
 
-/// A struct that steps through a ZeroTrie when fed data from fmt::Write
-struct ZeroTrieStepWrite<'a> {
-    cursor: ZeroTrieSimpleAsciiCursor<'a>,
-}
-
-impl<'a> fmt::Write for ZeroTrieStepWrite<'a> {
-    fn write_str(&mut self, s: &str) -> fmt::Result {
-        for b in s.bytes() {
-            if !b.is_ascii() {
-                return Err(fmt::Error);
-            }
-            self.cursor.step(b);
-        }
-        Ok(())
-    }
-    fn write_char(&mut self, c: char) -> fmt::Result {
-        if !c.is_ascii() {
-            return Err(fmt::Error);
-        }
-        self.cursor.step(c as u8);
-        Ok(())
-    }
-    fn write_fmt(&mut self, _: fmt::Arguments<'_>) -> fmt::Result {
-        unreachable!()
-    }
-}
-
 impl<'data> BlobSchemaV2<'data> {
     pub fn load(&self, key: DataKey, req: DataRequest) -> Result<&'data [u8], DataError> {
         let key_index = self
@@ -191,15 +163,12 @@ impl<'data> BlobSchemaV2<'data> {
             .locales
             .get(key_index)
             .ok_or_else(|| DataError::custom("Invalid blob bytes").with_req(key, req))?;
-        let mut trie_write = ZeroTrieStepWrite {
-            cursor: ZeroTrieSimpleAscii::from_store(zerotrie).into_cursor(),
-        };
+        let mut cursor = ZeroTrieSimpleAscii::from_store(zerotrie).into_cursor();
         #[allow(clippy::unwrap_used)] // infallible impl
         req.locale
-            .write_to(&mut trie_write)
+            .write_to(&mut cursor)
             .unwrap();
-        let blob_index = trie_write
-            .cursor
+        let blob_index = cursor
             .value()
             .ok_or_else(|| DataErrorKind::MissingLocale.with_req(key, req))?;
         let buffer = self
